Bring in MiscUtils for validation brevity.
Change "readonly" fields to properties with private setters.

Refactor the Template class so that it can have a ModifiedKey and do not
process the template until asked to. This is so that we can write a
TemplateLoader.

Add a "Templates" folder to the MSSQL plugin and start writing some
simple template files. Add custom build step to copy them into the
AutoZilla GUI's folder.